ROM: Eliminate ZIP_FORMAT
I have devised some logic to make the format that the ending zip is in irrelevant.

When a ROM completes, there are two scenarios: the zip is in the otapackage format, in which case there will be only one zip, or it will be in the bacon format, which will have the otapackage file and the bacon file. If there is one file, move it as is; if there is more than one, move only the ones that don't include the word ota

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>