funcs/cbl: Fix return code of kcbw
The exit code of a shell function is implicitly the return code of its
last command. To ensure that the build.sh exit code is properly passed
up, delete the last two commands. We don't need to reset hard anymore
because our out of tree modification is now in tree and the build config
is just a temp file, it will get reaped either by a manual cleanup or a
reboot.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>