env: Completely abstract setup to be done as root
This is a major refactoring of how setup is done. With this,
env/stubs/setup is sourced and run as root (be it with 'sudo -s' or
before an actual user account has been setup). This drastically
simplifies and unifies everything. Each supported architecture gets its
own setup function, where global setup commands are run (like updating
the distribution and installing packages). User specific setup will be
unified in a similar manner.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>