fish: cbl_qualify_tc_bld_uprev: Limit number of link jobs with LTO to avoid OOM
When building LLVM with LTO on a high core count system, it is possible
to run out of memory and have the whole job killed. There is likely no
heuristic that everyone would be happy with for limiting the number of
link jobs within build-llvm.py itself; instead, it can be done here via
the script's '--defines' argument.

Limit the number of jobs to 1 per 15GiB of system memory for ThinLTO
jobs and 1 per 30GiB of system memory for full LTO jobs.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>