fish: cbl_lkt: Harden against large log files
Now that llvm-kernel-testing adds QEMU boot logs to the log files, a
boot problem in a machine that constantly reboots could blow up the log
files. Gmail, which the notifier uses, has a 25MB limit to attachments,
so to avoid failures when sending files, write a small Python program
that adds the total size of all files and creates a tarball with them if
they are all larger than 25MB or prints all the paths to the non-zero
ones if not.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>