podman: dev/fedora: Temporarily downgrade to Fedora 38
Unfortunately, there is a regression in the kernel-headers package that
impacts my ability to compile kernels older than 6.5. While it may be
possible to hold back the kernel-headers package (although the dnf5
fiasco makes this harder), it is just easier to downgrade to Fedora 38,
which will still have pretty up to date tools.

Link: https://lore.kernel.org/linux-arch/20230727213648.GA354736@dev-arch.thelio-3990X/
Signed-off-by: Nathan Chancellor <nathan@kernel.org>