fish: git_setup: Avoid using '-X' with less
By default, git uses LESS=FRX but '-X' causes the output of the pager to
remain in the terminal window, which is generally undesirable. With
delta, use env to invoke it with LESS=FR. Without delta, use the '-+X'
flag to 'less' to override git's default environment.

If '-X' is genuinely desired, git can be invoked with

  $ git -c core.pager=$(if command -q delta; echo delta; else; echo less; end) ...

Signed-off-by: Nathan Chancellor <nathan@kernel.org>