python: setup: arch: Stop enabling reflector.service in VMs
While the idea of enabling Reflector for every boot seems like a good
idea in case VMs are not booted for some time, it is really disruptive
when doing quick build and boot debugging. Adjust the timer unit to
include a monotonic timer so that reflector runs if it has not been
active for at least 18 hours, which will avoid stomping the timers for
regular machines.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>