fish: variables: Set TMUX_TMPDIR to /var/tmp
In order to bind mound the tmux socket into a systemd-nspawn container
started automatically on boot, the file needs to exist but with the
socket on /tmp, it is only created when tmux is run for the first item,
which will be after the container needed to start. Use /var/tmp as the
base for the temporary tmux directory so that it persists across
reboots.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>