python: sd_nspawn: Add '--update'
Currently, there are only two ways to run a machine with the script:

  * through systemd-nspawn@.service via '--install'
  * temporarily via '--ephemeral'

With '--ephemeral', changes to the filesystem are not written back,
which means machines that are primarily run that way, such as
non-primary development containers, do not get updated unless they are
run as a service, which may not be desirable due to the frequency of
interaction with said container.

Add a new mode, '--update', which allows a user to drop into a
non-running machine as root to run the update commands for the
particular distribution.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>