fish: Dismantle run_cmd
Most of these commands will just be run from PATH. For the ones that
need some sort of customization, move that into their respective
function (even if it duplicates things) to make it clear what each one
needs and expects.

This kills '--tmux' with fzf, which has never worked since I
"introduced" it in commit c6e5bf8c ("fish: fzf: Switch to built in
'--tmux' flag") and it does not work easily with systemd-nspawn.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>