env: Update handling of mkosi virtual environment
A recent mkosi commit [1] necessitates updating the Buster patch applied
to support pgo-llvm-builder. With that change applied, only one line
needs to be adjusted to properly support Buster with the Mirror=
setting.

As part of this update, move mkosi handling into py_venv directly so
that it is easy to update and test mkosi outside of mkosi_bld.

Link: https://github.com/systemd/mkosi/commit/c5324fcdd26863c5349785f765fa34af0a45e394 [1]
Signed-off-by: Nathan Chancellor <nathan@kernel.org>